feat

+++ server/apps/monitor/services/monitor_object.py
@@ -260,7 +260,7 @@ class MonitorObjectService:
                 )
 
     @staticmethod
-    def update_instance(instance_id, name, organizations):
+    def update_instance(instance_id, name=None, organizations=None, **extra_fields):
         """更新监控对象实例"""
         instance = MonitorInstance.objects.filter(id=instance_id).first()
         if not instance:
@@ -268,12 +268,16 @@ class MonitorObjectService:
         if name:
             MonitorObjectService.validate_update_instance_name_unique(instance, name)
             instance.name = name
-            instance.save()
+        for field in ("cloud_region_id", "ip", "fallback_sampling_rate", "auto"):
+            if field in extra_fields and extra_fields[field] is not None:
+                setattr(instance, field, extra_fields[field])
+        instance.save()
 
         # 更新组织信息
-        instance.monitorinstanceorganization_set.all().delete()
-        for org in organizations:
-            instance.monitorinstanceorganization_set.create(organization=org)
+        if organizations is not None:
+            instance.monitorinstanceorganization_set.all().delete()
+            for org in organizations:
+                instance.monitorinstanceorganization_set.create(organization=org)
 
     @staticmethod
     def remove_instances_organizations(instance_ids, organizations):
@@ -298,8 +302,9 @@ class MonitorObjectService:
     @staticmethod
     def set_instances_organizations(instance_ids, organizations):
         """设置监控对象实例组织"""
-        if not instance_ids or not organizations:
+        if not instance_ids:
             return
+        organizations = organizations or []
 
         with transaction.atomic():
             # 删除旧的组织关联
@@ -310,4 +315,5 @@ class MonitorObjectService:
             for instance_id in instance_ids:
                 for org in organizations:
                     creates.append(MonitorInstanceOrganization(monitor_instance_id=instance_id, organization=org))
-            MonitorInstanceOrganization.objects.bulk_create(creates, ignore_conflicts=True)
+            if creates:
+                MonitorInstanceOrganization.objects.bulk_create(creates, ignore_conflicts=True)
